1.1.1 ACID Meet
* Test the commit procedure via comment on Main Function in consoleTest project.

* automation test

* test again (#2)


* Add comments for public methods.

* done

* optimize the write block size to 4096 bytes. (#10)

* Insert method returns the new id of the inserted data.

* find add a parameter to execute the search backward

* iterate without buffer whole page

* 1.000.002

* version edit

* fixed the initialize exception when directory dose not exist.

* check transaction

* a check for transaction dead lock in singl thread was added.

* Ver1.000.003

* fixed transaction conflicts in multi dbengine instance.

* add recursive  transaction supported

* 105

* remove read transaction exception method.

* ver

* move the lock outside the try-catch block to avoid repeat writelock deadlock.

* ver 1.0.7 ready

* Collected exception message

* timeout and exception message.

* fixed bug

* Make the destory process to be executed when disposing.
Remove Destory() and use SetDestoryOnDisposed() instead.

* done

* add dbEngine early dispose case check.

* add timeout parameter when safely disposed waiting for transaction finish.

* remove console write

* done

* done

* done

* 1_001_001

* readme Update 1.1.1

* version

---------

+++ LumDbEngine/Element/Engine/DbEngine.cs
@@ -4,6 +4,8 @@ using LumDbEngine.Element.Engine.Transaction;
 using LumDbEngine.Element.Exceptions;
 using LumDbEngine.Element.Structure;
 using LumDbEngine.IO;
+using System.Collections.Concurrent;
+using System.Diagnostics;
 
 namespace LumDbEngine.Element.Engine
 {
@@ -40,7 +42,7 @@ namespace LumDbEngine.Element.Engine
         public DbEngine(string path, bool createIfNotExists = true)
         {
             this.path = path;
-            InitializeEngine(this.path, createIfNotExists);
+              InitializeEngine(this.path, createIfNotExists);
         }
 
         private void InitializeEngine(string path, bool createIfNotExists)
@@ -56,34 +58,100 @@ namespace LumDbEngine.Element.Engine
                     throw new FileNotFoundException($"File not found:{path}");
                 }
             }
+
             this.iof = new IOFactory(path);
+            var state = DbLogUtils.CheckDbState(iof.RentReader());
+
+            switch (state)
+            {
+                case DbLogState.Writing:
+                    var dblog = DbLog.OpenLogToRecoveryDbEngine(this);
+                    dblog.DumpToDbEngine(iof.FileStream);
+                    dblog.Dispose();
+                    break;
+                case DbLogState.Done:
+                default:
+                    break;
+            }
+
         }
 
         private readonly AutoResetEvent autoResetEvent = new AutoResetEvent(true); // make sure serializable
 
-        internal void InitializeNew(string path)
+        private void InitializeNew(string path)
         {
-            var ck = new STChecker(autoResetEvent, callCount);
-            using var ts = new LumTransaction(null, ck, DbCache.DEFAULT_CACHE_PAGES, true);
-            ts.SaveAs(path);
-            ts.Discard();
+
+            try
+            {
+                var ck = new STChecker(autoResetEvent, callCount, -1);
+                using var ts = new LumTransaction(null, ck, DbCache.DEFAULT_CACHE_PAGES, true, this);
+                ts.SaveAs(path);
+                ts.Discard();
+            }
+            catch (Exception ex)
+            {
+                if (ex.Message == LumExceptionMessage.SingleThreadMultiTransaction || ex.Message == LumExceptionMessage.TransactionTimeout)
+                {
+                    throw;
+                }
+                else
+                {
+                    throw LumException.Raise("Transaction start failed, since DbEngine might be disposed.");
+                }
+            }
+
         }
 
 
         /// <summary>
-        /// Start a new transaction to execute commands. A started transaction must be disposed when action is done, and should better used with "using" in case of lacking diosposing. 
+        /// Start a new transaction to execute commands. A started transaction must be disposed when action is done, and should better use with "using" in case of not diosposing. 
         /// </summary>
         /// <param name="initialCachePages">Initial cache page size. The minimal page shoud be large than 128</param>
         /// <param name="dynamicCache">System manage the cache page automatically</param>
+        /// <param name="millisecondsTimeout">Milliseconds timeout waiting for different thread transaction done.</param>
         /// <returns></returns>
-        public ITransaction StartTransaction(int initialCachePages = DbCache.DEFAULT_CACHE_PAGES, bool dynamicCache = true)
+        public ITransaction StartTransaction(int initialCachePages = DbCache.DEFAULT_CACHE_PAGES, bool dynamicCache = true, int millisecondsTimeout = -1)
         {
-            var ck = new STChecker( autoResetEvent,callCount);
-            return new LumTransaction(iof, ck, initialCachePages, dynamicCache);
+
+            try
+            {
+
+                var ck = new STChecker(autoResetEvent, callCount, millisecondsTimeout);
+                return new LumTransaction(iof, ck, initialCachePages, dynamicCache, this);
+            }
+            catch (Exception ex)
+            {
+                if (ex.Message == LumExceptionMessage.SingleThreadMultiTransaction || ex.Message == LumExceptionMessage.TransactionTimeout)
+                {
+                    throw;
+                }
+                else
+                {
+                    throw LumException.Raise("Transaction start failed, since DbEngine might be disposed.");
+                }
+            }
+
         }
 
-        private bool disposed;
+        internal ConcurrentDictionary<Guid, ITransaction> transactionsPool = new();
 
+        /// <summary>
+        /// Get transaction id.
+        /// </summary>
+        /// <param name="id"></param>
+        /// <param name="ts"></param>
+        /// <returns></returns>
+        public bool GetById(Guid id, out ITransaction ts)
+        {
+            return transactionsPool.TryGetValue(id, out ts);
+        }
+
+        internal bool disposed;
+
+        /// <summary>
+        /// The milliseconds timeout when the dbEngine dispose waiting for the transaction end. Default value is 0, which means not waiting.
+        /// </summary>
+        public int DisposeMillisecondsTimeout { get; set; } = 0;
         /// <summary>
         /// Dispose the current engine and free the db file usage (if have).
         /// </summary>
@@ -91,23 +159,42 @@ namespace LumDbEngine.Element.Engine
         {
             if (disposed == false)
             {
-                disposed = true;
-                autoResetEvent?.Dispose();
+                if (DisposeMillisecondsTimeout > 0)
+                {
+                    autoResetEvent.WaitOne(DisposeMillisecondsTimeout);
+                    if (transactionsPool.Count > 0)
+                    {
+                        LumException.Throw($"{LumExceptionMessage.DbEngDisposedTimeOut} Living transactions: " +
+                            $"{string.Join(';', transactionsPool.Values.Select(o => o.Id.ToString()).ToArray())}");
+                    }
+                }
+
                 iof?.Dispose();
                 iof = null;
+                autoResetEvent?.Dispose();
+                disposed = true;
+                if (DesrotyOnDispose)
+                {
+                    if (File.Exists(path))
+                    {
+                        File.Delete(path);
+                    }
+                }
             }
         }
 
         /// <summary>
-        /// Physically delete the current db file on disk
+        /// Set desrotyOnDispose to be true and physically delete the current db file on disk when disposed.
         /// </summary>
-        public void Destory()
+        public void SetDestoryOnDisposed()
         {
-            Dispose();
-            if (File.Exists(path))
-            {
-                File.Delete(path);
-            }
+            DesrotyOnDispose = true;
         }
+
+        /// <summary>
+        /// Physically delete the current db file on disk when disposed.
+        /// </summary>
+        private bool DesrotyOnDispose { get; set; } = false;
+
     }
 }
